Create README.md
This is the personal website of Shifanth Jasim, a full-stack software engineer and freelance web developer. The site showcases my work, blog posts, and services, providing an interactive and professional hub for clients, collaborators, and those interested in my journey as a developer.

My website features a collection of my technical projects, each with detailed descriptions and links to live demos or repositories. Additionally, there is a blog section where I share programming insights, tutorials, and personal experiences in the world of coding. This site is designed using Bootstrap 5 to ensure a responsive and user-friendly experience across all devices.